first pass at changing trial conversion so researchstudy constructor is directly implemented by developer
