+++ README.md
@@ -10,15 +10,17 @@ These steps give an overview of steps you need to take to get the template into 
 1. Navigate to **env.ts**. 
     - Fill in the defaults environment variables corresponding to your api service
     - Create a file in the project directory named **.env.local** and enter the following: AUTH_TOKEN = **_insert authorization token_** 
-2. Open up **searchresponse.ts**. This file streamlines converting JSON responses from the API into actual objects.
-    - Complete the constructors for the _TrialObject_ and _SearchResponse_ classes
+2. Open up **searchset.ts**. This is one of two files that turns the response from the match API into a FHIR Bundle.
+    - Fill in the TO-DOs in the _SearchSet_ constructor
+3. Open up **research-study.ts**. This is the other file that turns the response from the match API into a FHIR Bundle. This FHIR maps a trial from the match API to a FHIR ResearchStudy.
+    - Complete the constructors for the _ResearchStudy_ classes
         * Extract as many paramaters as possible from the JSON while adhering to the spec
     - Ensure this part is completed as it used in the rest of the server 
-3. Open **query.ts**
+4. Open **query.ts**
     - Complete the _APIQuery_ class 
         * Learn more about patient bundle formatting on the [wiki](https://github.com/mcode/clinical-trial-matching-engine/wiki/Data-Model). 
         * Complete the **toQuery()** function which formats the request body sent to the matching service API. 
-4. Download and unzip folder of trials from (https://clinicaltrials.gov/ct2/resources/download#DownloadAllData) and place it in the src folder. Name it 'AllPublicXML'
+. Download and unzip folder of trials from (https://clinicaltrials.gov/ct2/resources/download#DownloadAllData) and place it in the src folder. Name it 'AllPublicXML'
 
 
 # Requirements
